api_test/requirements.txt
allure-pytest==2.8.33
allure-python-commons==2.8.33
Appium-Python-Client==1.0.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-order==0.9.2
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
